moved StatusMessage to shared model package
